+++ src/screens/StartupScreen/StartupScreen.test.tsx
